Move devolve function to weasel module
